requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
feedparser>=6.0.11
requests>=2.32.0
PyYAML>=6.0.1
python-dateutil>=2.9.0

# ── Database ──────────────────────────────────────────────────────────────────
SQLAlchemy>=2.0.0

# ── HTML parsing ──────────────────────────────────────────────────────────────
beautifulsoup4>=4.12.0
lxml>=5.0.0

# ── PDF generation ────────────────────────────────────────────────────────────
reportlab>=4.0.0

# ── Console output ────────────────────────────────────────────────────────────
rich>=13.0.0
